+++ R/dm-from-src.R
@@ -52,6 +52,12 @@ dm_from_src <- function(src = NULL, table_names = NULL, learn_keys = NULL,
     # FIXME: Check empty arguments and ellipsis
     return(empty_dm())
   }
+
+  if (inherits(src, "Pool")) {
+    src <- pool_src <- pool::poolCheckout(src)
+    on.exit(pool::poolReturn(pool_src))
+  }
+
   # both DBI-Connection and {dplyr}-src object are accepted
   src <- src_from_src_or_con(src)
   con <- con_from_src_or_con(src)
@@ -125,7 +131,6 @@ quote_ids <- function(x, con, schema = NULL) {
       ~ dbplyr::ident_q(dbplyr::build_sql(dbplyr::ident(.x), con = con))
     )
   } else {
-    if (!sql_schema_exists(con, schema)) abort_no_schema_exists(schema)
     map(
       x,
       ~ dbplyr::ident_q(schema_if(rep(schema, length(.x)), .x, con))