+++ R/error-helpers.R
@@ -143,7 +143,7 @@ abort_no_cycles <- function(g) {
 }
 
 error_txt_no_cycles <- function(shortest_cycle) {
-  c("Cycles in the relationship graph not yet supported.", glue::glue("Shortest cycle: {shortest_cycle}"))
+  c("Cycles in the relationship graph not yet supported.", i = glue::glue("Shortest cycle: {shortest_cycle}"))
 }
 
 
@@ -585,19 +585,6 @@ warn_if_arg_not <- function(arg,
 }
 
 # Errors for schema handling functions ------------------------------------
-
-abort_schema_exists <- function(schema, dbname = NULL) {
-  abort(error_txt_schema_exists(schema, dbname),
-    .subclass = dm_error_full("schema_exists")
-  )
-}
-
-error_txt_schema_exists <- function(schema, dbname) {
-  msg_suffix <- fix_msg(dbname)
-  glue(
-    "A schema named {tick(schema)} already exists{msg_suffix}."
-  )
-}
 
 abort_no_schema_exists <- function(schema, dbname = NULL) {
   abort(error_txt_no_schema_exists(schema, dbname),