+++ tests/testthat/test-schema.R
@@ -9,10 +9,10 @@ test_that("schema handling on MSSQL and Postgres works", {
     sql_schema_table_list_mssql
   }
 
-  expect_dm_error(sql_schema_exists(con_db, 1), "parameter_not_correct_class")
-  expect_dm_error(sql_schema_exists(con_db, letters[1:2]), "parameter_not_correct_length")
-  expect_dm_error(sql_schema_exists(src_db, 1), "parameter_not_correct_class")
-  expect_dm_error(sql_schema_exists(src_db, letters[1:2]), "parameter_not_correct_length")
+  expect_dm_error(db_schema_exists(con_db, 1), "parameter_not_correct_class")
+  expect_dm_error(db_schema_exists(con_db, letters[1:2]), "parameter_not_correct_length")
+  expect_dm_error(db_schema_exists(src_db, 1), "parameter_not_correct_class")
+  expect_dm_error(db_schema_exists(src_db, letters[1:2]), "parameter_not_correct_length")
 
   withr::defer({
     try(dbExecute(con_db, "DROP TABLE test_schema_1"))
@@ -20,28 +20,28 @@ test_that("schema handling on MSSQL and Postgres works", {
     try(dbExecute(con_db, SQL('DROP SCHEMA "1-dm_schema_TEST"')))
   })
 
-  expect_false(sql_schema_exists(con_db, "1-dm_schema_TEST"))
-  expect_false(sql_schema_exists(src_db, "1-dm_schema_TEST"))
+  expect_false(db_schema_exists(con_db, "1-dm_schema_TEST"))
+  expect_false(db_schema_exists(src_db, "1-dm_schema_TEST"))
 
   # create a table in the default schema
-  expect_message(sql_schema_create(con_db, "1-dm_schema_TEST"), "created")
-  expect_dm_error(sql_schema_create(con_db, "1-dm_schema_TEST"), "schema_exists")
+  expect_message(db_schema_create(con_db, "1-dm_schema_TEST"), "created")
+  expect_error(db_schema_create(con_db, "1-dm_schema_TEST"))
   expect_identical(
     con_db %>%
-      sql_schema_list(include_default = FALSE) %>%
+      db_schema_list(include_default = FALSE) %>%
       filter(schema_name == "1-dm_schema_TEST") %>%
       pull(schema_name),
     "1-dm_schema_TEST"
   )
-  expect_true(sql_schema_exists(con_db, "1-dm_schema_TEST"))
-  expect_message(sql_schema_drop(con_db, "1-dm_schema_TEST"), "Dropped schema")
-  expect_dm_error(sql_schema_drop(con_db, "1-dm_schema_TEST"), "no_schema_exists")
-  expect_false(sql_schema_exists(con_db, "1-dm_schema_TEST"))
+  expect_true(db_schema_exists(con_db, "1-dm_schema_TEST"))
+  expect_message(db_schema_drop(con_db, "1-dm_schema_TEST"), "Dropped schema")
+  expect_dm_error(db_schema_drop(con_db, "1-dm_schema_TEST"), "no_schema_exists")
+  expect_false(db_schema_exists(con_db, "1-dm_schema_TEST"))
 
-  expect_message(sql_schema_create(src_db, "1-dm_schema_TEST"), "created")
-  expect_true(sql_schema_exists(src_db, "1-dm_schema_TEST"))
-  expect_message(sql_schema_drop(src_db, "1-dm_schema_TEST"), "Dropped schema")
-  expect_false(sql_schema_exists(src_db, "1-dm_schema_TEST"))
+  expect_message(db_schema_create(src_db, "1-dm_schema_TEST"), "created")
+  expect_true(db_schema_exists(src_db, "1-dm_schema_TEST"))
+  expect_message(db_schema_drop(src_db, "1-dm_schema_TEST"), "Dropped schema")
+  expect_false(db_schema_exists(src_db, "1-dm_schema_TEST"))
 
   expect_false("test_schema_1" %in% sql_schema_table_list(con_db)$table_name)
   expect_false("test_schema_1" %in% sql_schema_table_list(src_db)$table_name)
@@ -66,7 +66,7 @@ test_that("schema handling on MSSQL and Postgres works", {
     tibble(a = 1:5)
   )
 
-  expect_message(sql_schema_create(src_db, "1-dm_schema_TEST"), "created")
+  expect_message(db_schema_create(src_db, "1-dm_schema_TEST"), "created")
 
   dbWriteTable(
     con_db,
@@ -74,19 +74,21 @@ test_that("schema handling on MSSQL and Postgres works", {
     value = tibble(b = letters[1:5])
   )
 
+  schema_list <- sql_schema_table_list(con_db, schema = "1-dm_schema_TEST")
+
   expect_identical(
-    sql_schema_table_list(con_db, schema = "1-dm_schema_TEST"),
+    schema_list,
     tibble(
       table_name = "test_schema_2",
       remote_name = dbplyr::ident_q("\"1-dm_schema_TEST\".\"test_schema_2\"")
     )
   )
 
-  remote_table_2 <- filter(
-    sql_schema_table_list(src_db, schema = "1-dm_schema_TEST"),
-    table_name == "test_schema_2"
-  ) %>%
+  remote_table_2 <-
+    schema_list %>%
+    filter(table_name == "test_schema_2") %>%
     pull(remote_name)
+
   expect_identical(
     tbl(src_db, remote_table_2) %>% collect(),
     tibble(b = letters[1:5])
@@ -99,7 +101,7 @@ test_that("schema handling on Postgres works", {
   src_db <- my_test_src()
   con_db <- src_db$con
 
-  sql_schema_create(con_db, "2-dm_schema_TEST")
+  db_schema_create(con_db, "2-dm_schema_TEST")
   dbWriteTable(
     con_db,
     DBI::Id(schema = "2-dm_schema_TEST", table = "test_schema_2"),
@@ -109,7 +111,7 @@ test_that("schema handling on Postgres works", {
   expect_identical(
     sort(
       filter(
-        sql_schema_list(con_db, include_default = TRUE),
+        db_schema_list(con_db, include_default = TRUE),
         schema_name == "public" | schema_name == "2-dm_schema_TEST"
       ) %>%
         pull(schema_name)
@@ -118,7 +120,7 @@ test_that("schema handling on Postgres works", {
   )
 
   expect_message(
-    sql_schema_drop(src_db, "2-dm_schema_TEST", force = TRUE),
+    db_schema_drop(src_db, "2-dm_schema_TEST", force = TRUE),
     "all objects"
   )
 })
@@ -135,36 +137,35 @@ test_that("schema handling on MSSQL works for different DBs", {
 
   original_dbname <- attributes(con_db)$info$dbname
   DBI::dbExecute(con_db, "CREATE DATABASE test_db_for_schema_dm")
-  expect_false(sql_schema_exists(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"))
+  expect_false(db_schema_exists(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"))
 
   expect_message(
-    sql_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
+    db_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
     "on database `test_db_for_schema_dm`"
   )
 
-  expect_dm_error(
-    sql_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
-    "schema_exists"
+  expect_error(
+    db_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm")
   )
 
   expect_identical(
     sort(
-      sql_schema_list(con_db, include_default = TRUE, dbname = "test_db_for_schema_dm")$schema_name
+      db_schema_list(con_db, include_default = TRUE, dbname = "test_db_for_schema_dm")$schema_name
     ),
     c("dbo", "test_schema")
   )
 
   expect_message(
-    sql_schema_drop(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
+    db_schema_drop(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
     "on database `test_db_for_schema_dm`"
   )
 
   expect_message(
-    sql_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
+    db_schema_create(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
     "on database `test_db_for_schema_dm`"
   )
 
-  expect_true(sql_schema_exists(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"))
+  expect_true(db_schema_exists(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"))
 
   expect_identical(
     sql_schema_table_list_mssql(con_db, schema = "test_schema", dbname = "test_db_for_schema_dm"),
@@ -182,7 +183,7 @@ test_that("schema handling on MSSQL works for different DBs", {
 
   expect_error(
     expect_warning(
-      sql_schema_drop(con_db, "test_schema", dbname = "test_db_for_schema_dm", force = TRUE),
+      db_schema_drop(con_db, "test_schema", dbname = "test_db_for_schema_dm", force = TRUE),
       "Argument `force` ignored:"
     )
   )
@@ -203,36 +204,36 @@ test_that("schema handling on SQLite all throw errors", {
   con_db <- src_db$con
 
   expect_dm_error(
-    sql_schema_exists(src_db, "test"),
+    db_schema_exists(src_db, "test"),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_list(src_db),
+    db_schema_list(src_db),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_drop(src_db, "test"),
+    db_schema_drop(src_db, "test"),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_exists(src_db, "test"),
+    db_schema_exists(src_db, "test"),
     "no_schemas_supported"
   )
 
   expect_dm_error(
-    sql_schema_exists(con_db, "test"),
+    db_schema_exists(con_db, "test"),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_list(con_db),
+    db_schema_list(con_db),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_drop(con_db, "test"),
+    db_schema_drop(con_db, "test"),
     "no_schemas_supported"
   )
   expect_dm_error(
-    sql_schema_exists(con_db, "test"),
+    db_schema_exists(con_db, "test"),
     "no_schemas_supported"
   )
 })