+++ NEWS.md
@@ -1,5 +1,58 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# dm 0.2.5
+
+## Features
+
+- `mutate()`, `transmute()`, `distinct()` and `summarize()` now support `dplyr::across()` and extra arguments (#640).
+- Key tracking for the first three verbs is less strict and based on name equality (#663).
+- `relocate()` now works on zoomed `dm` objects (#666).
+- `dm_add_fk()` gains `on_delete` argument which `copy_dm_to()` picks up and translates to an `ON DELETE CASCADE` or `ON DELETE NO ACTION` specification for the foreign key (#649).
+- `dm_copy_to()` defines foreign keys during table creation, for all databases except DuckDB. Tables are created in topological order (#658). For cyclic relationship graphs, table creation is attempted in the original order and may fail (#664).
+- `waldo::compare()` shows better output for dm objects (#642).
+- `dm_paste()` output uses trailing commas in the `dm::dm()` and `tibble::tibble()` calls, and sorts column attributes by name, for better modularity (#641).
+
+## Breaking changes
+
+- New `db_schema_create()`, `db_schema_drop()`, `db_schema_exists()` and `db_schema_list()` replace the corresponding `sql_schema_*()` functions, the latter are soft-deprecated (#670). The connection argument to `db_schema_*()` is called `con`, not `dest` (#668).
+
+## Bug fixes
+
+- `copy_dm_to()` and `sql_create_schema()` no longer actively check for schema existence (#644, #660).
+- Add newline after `OUTPUT` clause for SQL Server (#647).
+- Fix `sql_rows_delete()` with `returning` argument for SQL Server (#645).
+
+## Internal
+
+- Remove method only needed for RSQLite < 2.2.8, add warning if loaded RSQLite version is <= 2.2.8 (#632).
+- Adapt MSSQL tests to testthat update (#648).
+
+
+# dm 0.2.4
+
+## Features
+
+- `rows_insert()`, `rows_update()` and `rows_delete()` gain `returning` argument. In combination with `in_place = TRUE` this argument makes the newly inserted rows accessible via `get_returning_rows()` after the operation completes (#593, @mgirlich).
+- Implement `rows_patch()` for DBI connections (#610, @mgirlich).
+- Use `NO ACTION` instead of `CASCADE` in foreign key constraints to permit self-references.
+- `dm_from_src()` supports `pool::Pool` objects (#599, @moodymudskipper).
+- Better error message for `dm_rows_update()` and related functions for dm objects with tables without primary key (#592).
+- `glimpse()` is implemented for `dm` objects (#605).
+- Support DuckDB in `rows_insert()`, `rows_update()` and `rows_delete()` (#617, @mgirlich).
+
+## Bug fixes
+
+- Fix `dm_zoom_to()` for `dm` objects with an empty table (#626, @moodymudskipper).
+- Avoid generating invalid `dm` objects in some corner cases (#596).
+
+## Internal
+
+- `sql_schema_list()` supports `pool::Pool` objects (#633, @brancengregory).
+- Establish compatibility with pillar 1.6.2, vctrs > 0.3.8 and rlang > 0.4.11 (#613).
+- Use `check_suggested()` everywhere (#572, @moodymudskipper).
+- Add CI run for validating all new `dm` objects (#597).
+
+
 # dm 0.2.3
 
 ## Bug fixes